[FEATURE] implements post editing & post history
Posts become now historical posts when they are edited, and
a new post is generated and assigned to the thread. Historical
posts don't have a thread. Because crdate is defined by the system,
publishingDate is introduced to store the creation date of the
original post, even if the post is not the original post, so it
is possible that the crdate of a post is later than its publishing
date. However, if this is the case, the post must have at least
one historical post attached.